Move print functions to own module inside nu-engine
